borrow book  done

+++ src/app/modules/Books/book.service.ts
@@ -1,6 +1,8 @@
 import prisma from "../../../shared/prisma";
 import { Prisma } from "@prisma/client";
 import { TBook, TBookUpdate } from "./Book.interface";
+import ApiError from "../../errors/ApiError";
+import httpStatus from "http-status";
 
 // Define createBook function to insert a new book into the database
 const createBook = async (payload: Prisma.BookCreateInput): Promise<TBook> => {
@@ -12,6 +14,7 @@ const createBook = async (payload: Prisma.BookCreateInput): Promise<TBook> => {
 
 const getAllBooks = async () => {
   const books = await prisma.book.findMany();
+
   return books;
 };
 
@@ -19,6 +22,9 @@ const getSingleBooks = async (bookId: string) => {
   const book = await prisma.book.findUnique({
     where: { bookId: bookId },
   });
+  if (!book) {
+    return new ApiError(httpStatus.NOT_FOUND, "Book not found");
+  }
   return book;
 };
 
@@ -27,6 +33,7 @@ const updateBook = async (
   payload: Partial<TBookUpdate>
 ): Promise<TBook> => {
   const { title, genre, publishedYear, totalCopies, availableCopies } = payload;
+
   const updatedBook = await prisma.book.update({
     where: { bookId: bookId },
     data: {
@@ -41,12 +48,21 @@ const updateBook = async (
 };
 
 const deleteBook = async (bookId: string) => {
+  const book = await prisma.book.findUnique({
+    where: { bookId: bookId },
+  });
+  if (!book) {
+    return new ApiError(httpStatus.NOT_FOUND, "Book not found");
+  }
   const result = await prisma.book.delete({
     where: { bookId: bookId },
   });
 
   return result;
 };
+
+
+
 
 export const BookService = {
   createBook,
